Implemented event queuing.
This project now utilizes RabbitMQ to queue events for processing.  Submitting an incident to ServiceNow can sometimes take a bit of time.  In the previous version, if your were receiving a spike in syslog events, the script wouldn't potentially be too busy to handle them.  This new version will queue events as they're recieved, and the event publisher will process each one.